Moved OpmLog from opm-parser to opm-common
