--- opm/parser/eclipse/OpmLog/tests/CMakeLists.txt
@@ -1,3 +1,0 @@
-opm_add_test(runOpmLogTests SOURCES OpmLogTests.cpp
-                            EXE_NAME runOpmLogTests
-                            LIBRARIES opmparser ${Boost_LIBRARIES})
