+++ opm/parser/eclipse/Parser/Parser.cpp
@@ -26,8 +26,8 @@
 
 #include <opm/json/JsonObject.hpp>
 
-#include <opm/parser/eclipse/OpmLog/LogUtil.hpp>
-#include <opm/parser/eclipse/OpmLog/OpmLog.hpp>
+#include <opm/common/OpmLog/LogUtil.hpp>
+#include <opm/common/OpmLog/OpmLog.hpp>
 
 #include <opm/parser/eclipse/Deck/Deck.hpp>
 #include <opm/parser/eclipse/Deck/DeckItem.hpp>
